Revert plist real to float
Double value should be the correct type given the specification for the
binary plist format, however, we have a rounding issue when converting
to String in XML and ASCII plist format. The recommendation should be to
use BigDecimal but for now, we will just use float.

+++ subprojects/xcode-ide-kit/src/main/java/dev/nokee/xcode/AsciiPropertyListReader.java
@@ -255,7 +255,7 @@ public final class AsciiPropertyListReader implements PropertyListReader {
 	}
 
 	@Override
-	public double readReal() {
+	public float readReal() {
 		throw new UnsupportedOperationException();
 	}
